Extract common test helpers to own files (#336)
Proxy test file is overused of multiple logics.
Make it a bit cleaner, moved `WithTCPServer` to testhelper.
Tried to move `WithTCPProxy`, but current structure does not allow it.
If move it under package `testhelper`, it introduces import packages loop.
For now extract it to separate helper: `toxiproxy_test.go`.

In future it would nice to move business logic code under folder `pkg`.
