+++ testhelper/timeout_after_test.go
@@ -1,17 +1,19 @@
-package testhelper
+package testhelper_test
 
 import (
 	"testing"
 	"time"
+
+	"github.com/Shopify/toxiproxy/v2/testhelper"
 )
 
 func TestTimeoutAfter(t *testing.T) {
-	err := TimeoutAfter(5*time.Millisecond, func() {})
+	err := testhelper.TimeoutAfter(5*time.Millisecond, func() {})
 	if err != nil {
 		t.Fatal("Non blocking function should not timeout.")
 	}
 
-	err = TimeoutAfter(5*time.Millisecond, func() {
+	err = testhelper.TimeoutAfter(5*time.Millisecond, func() {
 		time.Sleep(time.Second)
 	})
 	if err == nil {
